[Refactor] : bookmark 패키지 분리
