+++ core/src/main/java/dev/handsup/bookmark/domain/Bookmark.java
@@ -1,10 +1,11 @@
-package dev.handsup.auction.domain;
+package dev.handsup.bookmark.domain;
 
 import static jakarta.persistence.ConstraintMode.*;
 import static jakarta.persistence.FetchType.*;
 import static jakarta.persistence.GenerationType.*;
 import static lombok.AccessLevel.*;
 
+import dev.handsup.auction.domain.Auction;
 import dev.handsup.common.entity.TimeBaseEntity;
 import dev.handsup.user.domain.User;
 import jakarta.persistence.Column;
