+++ core/src/main/java/dev/handsup/bookmark/dto/FindUserBookmarkResponse.java
@@ -1,4 +1,4 @@
-package dev.handsup.auction.dto.response;
+package dev.handsup.bookmark.dto;
 
 import java.time.LocalDateTime;
