+++ core/src/test/java/dev/handsup/auction/repository/auction/AuctionRepositoryTest.java
@@ -13,9 +13,10 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Slice;
 
 import dev.handsup.auction.domain.Auction;
-import dev.handsup.auction.domain.Bookmark;
+import dev.handsup.bookmark.domain.Bookmark;
 import dev.handsup.auction.domain.product.product_category.ProductCategory;
 import dev.handsup.auction.repository.product.ProductCategoryRepository;
+import dev.handsup.bookmark.repository.BookmarkRepository;
 import dev.handsup.fixture.AuctionFixture;
 import dev.handsup.fixture.BookmarkFixture;
 import dev.handsup.fixture.ProductFixture;
